+++ RFAlloctionProxy/RFAlloctionProxyButton.m
@@ -1,0 +1,7 @@
+
+#import "RFAlloctionProxyButton.h"
+#import "RFPerformance.h"
+
+@implementation RFAlloctionProxyButton
+
+@end