+++ RFGeometry.h
@@ -1,9 +1,10 @@
 /*!
     RFGeometry
     RFKit
+ 
     Math and Graphics helper
  
-    Copyright (c) 2012 BB9z
+    Copyright (c) 2012-2013 BB9z
     http://github.com/bb9z/RFKit
  
     The MIT License (MIT)
@@ -14,6 +15,10 @@
 
 #ifndef _RFKit_RFGeometry_h_
 #define _RFKit_RFGeometry_h_
+
+#ifndef RFMathNotChange
+#   define RFMathNotChange CGFLOAT_MAX
+#endif
 
 typedef enum {
 	RFResizeAnchorCenter = 0,