+++ category/Foundation/NSFileManager+RFKit.h
@@ -1,8 +1,8 @@
 /*!
-    RFKit
     NSFileManager extension
+    RFKit
 
-    Copyright (c) 2012 BB9z
+    Copyright (c) 2012-2013 BB9z
     http://github.com/bb9z/RFKit
 
     The MIT License (MIT)
@@ -14,7 +14,9 @@
 @interface NSFileManager (RFKit)
 - (NSArray *)subDirectoryOfDirectoryAtPath:(NSString *)path error:(NSError **)erro;
 
-- (unsigned long long)fileSizeForPath:(NSString *)path;
-
+/** Get file size of the given path.
+    @return file size. If the file not exist, the return value will be 0. If error occurs, the return value will be -1.
+ */
+- (long long)fileSizeForPath:(NSString *)path;
 
 @end