+++ category/UIKit/UIButton+RFKit.h
@@ -1,8 +1,8 @@
 /*!
-    RFKit
     UIButton extension
+    RFKit
 
-    Copyright (c) 2012 BB9z
+    Copyright (c) 2012-2013 BB9z
     http://github.com/bb9z/RFKit
 
     The MIT License (MIT)
@@ -12,7 +12,6 @@
 #import "RFRuntime.h"
 
 @interface UIButton (RFKit)
-
 
 /**
  Give a image name, this will auto set a button`s image of all state.