+++ category/UIKit/UIImage+RFKit.h
@@ -1,8 +1,8 @@
 /*!
-    RFKit
     UIImage extension
+    RFKit
 
-    Copyright (c) 2012 BB9z
+    Copyright (c) 2012-2013 BB9z
     http://github.com/bb9z/RFKit
 
     The MIT License (MIT)