+++ category/UIKit/UIView+RFKit.m
@@ -24,6 +24,24 @@
     return img;
 }
 
++ (void)animateWithDuration:(NSTimeInterval)duration delay:(NSTimeInterval)delay options:(UIViewAnimationOptions)options animated:(BOOL)animated beforeAnimations:(void (^)(void))before animations:(void (^)(void))animations completion:(void (^)(BOOL finished))completion {
+    if (animated) {
+        if (before) {
+            before();
+        }
+        [UIView animateWithDuration:duration delay:delay options:options animations:animations completion:completion];
+    }
+    else {
+        if (animations) {
+            animations();
+        }
+        if (completion) {
+            completion(YES);
+        }
+    }
+}
+
+
 #pragma mark 视图位置／尺寸
 
 -(void)exhangeWidthHight {
@@ -31,6 +49,8 @@
 	self.frame = CGRectMake(tmp.origin.x, tmp.origin.y, tmp.size.height, tmp.size.width);
 }
 
+#pragma clang diagnostic push
+#pragma clang diagnostic ignored -Wfloat-equal
 - (void)moveX:(CGFloat)x Y:(CGFloat)y {
 	CGPoint tmp = self.center;
 	if (x != RFMathNotChange) tmp.x += x;
@@ -55,6 +75,7 @@
     CGRect targetFrame = CGRectResize(self.frame, newSize, resizeAnchor);
     self.frame = targetFrame;
 }
+#pragma clang diagnostic pop
 
 - (CGFloat)distanceBetweenFrameBottomAndSuperviewBottom {
 	CGRect frame = self.frame;
@@ -120,10 +141,9 @@
 			
 		case RFViewResizeOptionCenter:
 			view.center = self.center;
-			return;
 			break;
-			
-		default:
+            
+        case RFViewResizeOptionNone:
 			break;
 	}
 }
@@ -133,7 +153,7 @@
         [view removeFromSuperview];
     }
     else {
-        douts(@"Warning: RFKit [UIView removeSubview] 父视图没有指定的子视图")
+        dout_warning(@"RFKit [UIView removeSubview] 父视图没有指定的子视图")
     }
 }
 
@@ -188,11 +208,11 @@
 }
 
 -(BOOL)isInFront{
-	return ([self.superview.subviews lastObject]==self);
+	return ([self.superview.subviews lastObject] == self);
 }
 
 -(BOOL)isAtBack{
-	return ([self.superview.subviews objectAtIndex:0]==self);
+	return ([self.superview.subviews objectAtIndex:0] == self);
 }
 
 -(void)exchangeDepthsWithView:(UIView*)swapView{