+++ ChangeLog.md
@@ -1,6 +1,15 @@
 RFKit Change Log
 ================
 
+1.1
+-----
+* Hello 2013~
+* Eliminate warnings.
+* dout 1.3.0.
+* New NSJSONSerialization, UIWebView extensions.
+* @keypath.
+* Document improve.
+
 1.0
 -----
 * Add Chinese document.